fix unrecognized print format specifier for int8/uint8 (#1303)
* fix unrecognized print format specifier for int8/uint8

* use c++ static_cast instead of c cast style

+++ include/cute/util/print.hpp
@@ -70,13 +70,13 @@ print(char c) {
 CUTE_HOST_DEVICE
 void
 print(signed char a) {
-  printf("%hhd", a);
+  printf("%d", static_cast<int>(a));
 }
 
 CUTE_HOST_DEVICE
 void
 print(unsigned char a) {
-  printf("%hhu", a);
+  printf("%u", static_cast<unsigned int>(a));
 }
 
 CUTE_HOST_DEVICE
